// glassdoor/src/app/components/employers/users/models/register_model.ts
export class Register {
    // firstName: String;
    // lastName: String;
    // username: String;
    // email: String;
    // password: String;

    firstName: String;
    lastName: String;
    workEmail: String;
    jobOpening: String;
    password: String;
    company: String;
    jobTitle: String
}